Rename random to fluky to avoid name conflict, fix typos and address inspection results
